refactor tools into better dir structure
