+++ packages/mcp-lexicon/src/coding/tools/bash/description.md
