+++ packages/mcp-lexicon/src/coding/tools/bash/mod.rs
