+++ packages/mcp-lexicon/src/coding/tools/bash/read_background_description.md
