+++ packages/mcp-lexicon/src/coding/tools/find/description.md
