+++ packages/mcp-lexicon/src/coding/tools/grep/common.rs
