+++ packages/mcp-lexicon/src/coding/tools/grep/description.md
