+++ packages/mcp-lexicon/src/coding/tools/lsp/diagnostics/description.md
