+++ packages/mcp-lexicon/src/coding/tools/lsp/find_references/description.md
