+++ packages/mcp-lexicon/src/coding/tools/lsp/goto_definition/description.md
