+++ packages/mcp-lexicon/src/coding/tools/lsp/hover/description.md
