+++ packages/mcp-lexicon/src/coding/tools/lsp/hover/mod.rs
@@ -4,7 +4,7 @@ use lsp_types::Hover;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use super::common::{LocationResult, parse_line};
+use crate::coding::lsp::common::{LocationResult, parse_line};
 use crate::coding::tools_trait::CodingTools;
 
 /// Input for the lsp_hover tool
