+++ packages/mcp-lexicon/src/coding/tools/lsp/mod.rs
@@ -1,0 +1,17 @@
+//! LSP tools module
+//!
+//! This module contains LSP-powered tools and the LspCodingTools wrapper.
+
+pub mod coding_tools;
+pub mod diagnostics;
+pub mod find_references;
+pub mod goto_definition;
+pub mod hover;
+pub mod tool;
+pub mod workspace_symbol;
+
+// Re-export the main LspCodingTools wrapper
+pub use coding_tools::LspCodingTools;
+
+// Re-export all tool types through the tool module
+pub use tool::*;
