+++ packages/mcp-lexicon/src/coding/tools/lsp/tool.rs
@@ -1,0 +1,19 @@
+//! LSP tools for querying language server information
+//!
+//! This module re-exports LSP tool types and functions from the individual tool modules.
+
+pub use super::diagnostics::{
+    DiagnosticsSummary, LspDiagnostic, LspDiagnosticsInput, LspDiagnosticsOutput,
+    execute_lsp_diagnostics,
+};
+pub use super::find_references::{
+    LspFindReferencesInput, LspFindReferencesOutput, execute_lsp_find_references,
+};
+pub use super::goto_definition::{
+    LspGotoDefinitionInput, LspGotoDefinitionOutput, execute_lsp_goto_definition,
+};
+pub use super::hover::{LspHoverInput, LspHoverOutput, execute_lsp_hover};
+pub use super::workspace_symbol::{
+    LspWorkspaceSymbolInput, LspWorkspaceSymbolOutput, SymbolResult, execute_lsp_workspace_symbol,
+};
+pub use crate::coding::lsp::common::LocationResult;
