+++ packages/mcp-lexicon/src/coding/tools/lsp/workspace_symbol/description.md
