+++ packages/mcp-lexicon/src/coding/tools/read_file/description.md
