+++ packages/mcp-lexicon/src/coding/tools/write_file/description.md
